requirements.txt
torch==2.2.0
torchvision==0.17.0 
torchaudio==2.2.0
matplotlib==3.8.2
numpy==1.26.3
Pillow==10.2.0
wandb==0.13.5
fiona==1.9.5
rasterio==1.3.9
shapely==2.0.2
imutils==0.5.4
opencv-contrib-python==4.9.0.80
tqdm==4.66.1
scikit-learn==1.4.0
